refactor: consolidate config package from 5 files to 2+init
  Merge config.py, settings.py, and mixins.py into models.py and __init__.py.
  No behavioral changes.
